syntree
Trait Flavor

pub trait Flavor {
    type Error;
    type Index: Index<Length = Self::Length>;
    type Length: Length;
    type Width: Width<Pointer = Self::Pointer>;
    type Pointer: Pointer;
    type Storage<T>: Storage<T, Error = Self::Error>;
    type Indexes: Storage<TreeIndex<Self>, Error = Self::Error>;
}

The flavor of a tree.

This should not be implemented directly, instead see the flavor! macro.

The Index associated type is constrained by the Index trait, and determines the numerical bounds of spans the tree.

The Width associated type determines the bounds of pointers in the tree through the Width trait, this decides how many elements that can be stored in the tree.

Required Associated Types

type Error

The error raised by the type of the tree.

type Index: Index<Length = Self::Length>

The type of an index used by a tree.

type Length: Length

The length used in the flavor.

type Width: Width<Pointer = Self::Pointer>

The width used in the flavor.

type Pointer: Pointer

The pointer in use.

type Storage<T>: Storage<T, Error = Self::Error>

The storage type used in the tree.

type Indexes: Storage<TreeIndex<Self>, Error = Self::Error>

How indexes are stored in the tree.

Object Safety

This trait is not object safe.