Trait syntree::Index

pub trait Index:
    Sized
    + Copy
    + Ord
    + Eq
    + Sealed { }

A type that can be used when referring to an index in a tree.

An index is a valid single component of a Span, valid indexes are types such as u32 and usize, but also Empty in case indexing is not required.

See Builder::new_with.

Object Safety

This trait is not object safe.