Module syntree::pointer

Pointer-related types and traits.

Structs

PointerU8

Pointer implementation for u8.

PointerU16

Pointer implementation for u16.

PointerU32

Pointer implementation for u32.

PointerU64

Pointer implementation for u64.

PointerU128

Pointer implementation for u128.

PointerUsize

Pointer implementation for usize.

Traits

Pointer

A pointer type that is derived from the pointer Width.

Width

A pointer width that can be used to reference other nodes.