syntree::pointer

Trait Pointer

pub trait Pointer:
    Sized
    + Copy
    + Hash
    + Eq
    + Debug
    + Sealed {
    // Required methods
    fn new(value: usize) -> Option<Self>;
    fn get(self) -> usize;
}

A pointer type that is derived from the pointer Width.

Required Methods

fn new(value: usize) -> Option<Self>

Safe constructor for the pointer.

Returns None if the value provided is out of bounds for the pointer type.

Examples

use syntree::pointer::{Pointer, PointerU32};

let v = PointerU32::new(0).ok_or("out of bounds")?;
assert_eq!(v.get(), 0);
let v = PointerU32::new(42).ok_or("out of bounds")?;
assert_eq!(v.get(), 42);

fn get(self) -> usize

Get the index of a pointer.

Examples

use syntree::pointer::{Pointer, PointerU32};

let v = PointerU32::new(0).ok_or("out of bounds")?;
assert_eq!(v.get(), 0);
let v = PointerU32::new(42).ok_or("out of bounds")?;
assert_eq!(v.get(), 42);

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.